% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fnAve}
\alias{fnAve}
\title{Helper function for obtaining the average.}
\usage{
fnAve(x)
}
\arguments{
\item{x}{Vector of data.}
}
\value{
Average value.
}
\description{
This function computes the average.
}

