% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fnFreqDay}
\alias{fnFreqDay}
\title{Helper function for obtaining the most frequent day of purchase.}
\usage{
fnFreqDay(x)
}
\arguments{
\item{x}{Vector of data.}
}
\value{
Most frequent day of purchase.
}
\description{
This function computes the most frequent day of purchase.
}

