% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fnMax}
\alias{fnMax}
\title{Helper function for obtaining the maximum.}
\usage{
fnMax(x)
}
\arguments{
\item{x}{Vector of data.}
}
\value{
Maximum value.
}
\description{
This function computes the maximum.
}

