% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fnSum}
\alias{fnSum}
\title{Helper function for obtaining the sum.}
\usage{
fnSum(x)
}
\arguments{
\item{x}{Vector of data.}
}
\value{
Sum value.
}
\description{
This function computes the sum.
}

