% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosted.R
\name{incrementality_func}
\alias{incrementality_func}
\title{Statistic for the Bootstrap}
\usage{
incrementality_func(datadiff, indices)
}
\arguments{
\item{datadiff}{Differences of the chosen metric}

\item{indices}{The indices of the sampled rows}
}
\value{
Bootstrap mean and standard error
}
\description{
This function defines the statistic for the boostrap.
}

