% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosted.R
\name{incrementality_func_hypothesis}
\alias{incrementality_func_hypothesis}
\title{Hypothesis Testing Bootstrap}
\usage{
incrementality_func_hypothesis(datadiff, indices)
}
\arguments{
\item{datadiff}{Differences of the chosen metric centered from the mean of the difference}

\item{indices}{The indices of the sampled rows}
}
\value{
Bootstrap mean and standard error
}
\description{
This function construct the sampling distribution of a metric under a null effect.
}

