% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{incrementality_metrics}
\alias{incrementality_metrics}
\title{Function to calculate RPU, AOV, BR, and TPB}
\usage{
incrementality_metrics(nb_transactions, nb_users, revenue, nb_buyers)
}
\arguments{
\item{nb_transactions}{The number of transactions}

\item{nb_users}{The number of unique users}

\item{revenue}{The revenue}

\item{nb_buyers}{The number of buyers}
}
\value{
RPU The Revenue per user

AOV The Average order value

BR The Buyer rate

TPB The Transactions per buyer
}
\description{
This function will calculate the four metrics commonly use in incrementality testing.
}
\examples{
incrementality_metrics(nb_transactions = 11278, nb_users=297073, revenue= 279480.4, nb_buyers=10909)
}

