% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boosted.R
\name{res_boot}
\alias{res_boot}
\title{Boostrap routine}
\usage{
res_boot(mydatadiff, statisticfunc, nb_boot)
}
\arguments{
\item{mydatadiff}{Contains the differences}

\item{statisticfunc}{Statistic to be bootstrapped}

\item{nb_boot}{Number of bootstrap samples}
}
\value{
Bootstrap object from package "boot"
}
\description{
This function runs the bootstrap routine.
}

