% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_confidence.R
\name{t_test_cf}
\alias{t_test_cf}
\title{Function to calculate Student-t confidence interval}
\usage{
t_test_cf(datawithID, confinter)
}
\value{
Student-t confidence interval
}
\description{
This function will calculate the Student-t confidence interval of a given sample of mean differences.
}

