% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_metric.R
\name{test_metric}
\alias{test_metric}
\title{Construct a dataframe containing the ab test id and the differences of the specified metric}
\usage{
test_metric(mydata, metric)
}
\arguments{
\item{mydata}{A dataframe containg the  variables "iabtest_id", "abt_group", and the metric of choice.}

\item{metric}{Chosen metric e.g. RPU, AOV, TPB, and BR}
}
\value{
Dataframe contaning the differences of the specified metrics for the control and test across all ab tests
}
\description{
This function caculates the differences of the specified metrics (e.g. RPU, AOV, etc.).
}

